Random Forest P-Score Model
Used Lending Club data and filtered for only small business data to create a
Random Forest classifier Model.

The user inputs the necessary information about the small business found in data.json.
The model will output a P-Score calculated by finding the probability of the test
sample being classified as "Not Default" (1).

Model also looks for business on Yelp, and factors in Yelp rating if business is found.